refactor: extract ref action recovery helpers
